Made Set-SFVMwareBestPractices an Advanced Function
Increased capabilities of Set-SFVMwareBestPractices to make it simpler
to implement all or part of the configurations as well as ability to set
values for some of the configurations different then the recommended.
This will allow ability to make modifications more specific to customer
environments in the event that it differs slightly from recommended
settings. Similar work and feel as the Get-SFVMwareBestPractices
function.