Additional Logic
Addition of new logic:
Check if iSCSI HBA adapter already has the SolidFire system SVIP as a
send target before trying to add it which would result in an error.
Check if iSCSI HBA IQN already exists in Volume Access Group before
trying to add it which would result in an error.